Added impl for image_view_any and fix to benchmark that was missing include for mapnik/color.hpp
